Major refactoring for Java 11
Major code refactoring with several goals:
* Compatibility with Java 11, including modularization
* Consolidation of processing code (i.e. no more ImageJ/OpenCV decision; permit both to be used together, with other key libraries)
* Increase the chances of finding a suitable packaging solution to create cross-platform installers

There remain many limits... e.g. modularization is incomplete. because it was incompatible with some dependencies (e.g. OpenCV with JavaCPP); the biggest challenge is the need for multiple modules not to use the same package names.

This also resulted in some QuPath classes being moved, and in general most packages are exposed in all modules... but this is probably not desirable.  So the next phase needs to look at improving the organization further.