Revised pressure sensitive input
Simplified JPen extension and introduced pressure-sensitivity with the Wand tool.

Attempted to work around a bug where activating the pen would then break the normal brush/wand when used with a mouse. Now if pressure and tilt values are all 0 it is assumed to mean we don't have a pen after all, and should use 'full pressure' (i.e. a standard mouse).